pyproject.toml
[tool.poetry]
name = "ingest"
version = "0.1.0"
description = ""
authors = ["Patrik Persson <patrik95195@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
azure-mgmt-datafactory = "^7.0.0"
azure-identity = "^1.16.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
